Fix --flatten flag parsing
Lines starting with `--` are considered options in docopt, and
`--flatten` was present at the start of a description line as
well as listed in the options sectioon; this caused docopt to
record --flatten twice, giving the cryptic message:

`--flatten is not a unique prefix: --flatten, --flatten?`

sigh.

docopt does have some serious downsides.