Add grammar support for a primary key statement
Primary key statements allow one to configure the name and
value generation for the primary key. The statement may be
used at the top level to denote the primary key for all
entities, and may also be present at the top of entity
bodies to override the primary key for a particular entity
expression.

NOTE: This is a grammar change only; the parser will parse the
statements, but they are essentially no-op until we write
interpreter and generator code to handle these statements.