Fix a bug in deferred evaluation
When handling assignment and variable declaration deferred evaluation,
the DeferredResolver would incorrectly cache the final value after the
first run, causing variables to hold the same value in subsequent
invocations of an assignment or variable declaration DeferredResolver.

This affected lambda closures that updated the closed variable within
the lambda body; essentially, the closed variable's value would never
change on invocations subsequent to the first.

What a pain in the ass to figure this one out.