Prevent variables, when declared in entity field expressions, from being added as field entries in the entity result
As the entity result doubles as the symbol table for the transient scope,
let, lambda, and entity expressions would inadvertantly declare fields in
the entity result, which is undesirable (especially being that Lambda and
Generator types will not serialize to JSON).

Creating another child scope from the transient scope ensures that any
symbols introduced by evaluating expressions will be inserted into a new
symbol table, effectively protecting the entity result. Note that it is
still possible to clobber a field via assignment expression, but this
is expected as fields are basically normal variables in this context.