Add $uniqint(), and allow $incr() to take an option offset argument.
The combination of $uniqint(), $incr(offset), and $uid() should give
the user the ability to uniquify just about anything.

Ensuring $dict() and $enum() to be unique may not be all that useful
in practice; $enum() is most likely going to be used for smaller,
finite sets that allow repetition (we believe, anyhow).

$dict() values (and other types) can be uniqified by concatenating
with $uniqint(), $incr(), or $uid().